+++ x-pack/platform/plugins/shared/agent_builder/public/application/components/conversations/conversation_input/message_editor/use_message_editor.ts
@@ -11,7 +11,13 @@ import type { CommandMatchResult, CommandBadgeData } from './command_menu';
 import { useCommandMenu, useCommandMenuPrefetch } from './command_menu';
 import { createCommandBadgeElement, deserializeCommandBadge } from './command_badge';
 import { serializeEditorContent } from './serialize';
-import { createCommandRange, insertSpaceAfter, placeCursorAfter, placeCursorAtEnd } from './utils';
+import {
+  createCommandRange,
+  createTextFragment,
+  insertSpaceAfter,
+  placeCursorAfter,
+  placeCursorAtEnd,
+} from './utils';
 
 export interface MessageEditorInstance {
   ref: React.RefObject<HTMLDivElement>;
@@ -154,15 +160,7 @@ const useMessageEditorController = ({
 
         for (const segment of segments) {
           if (segment.type === 'text') {
-            const parts = segment.value.split('\n');
-            parts.forEach((part, i) => {
-              if (part) {
-                ref.current!.appendChild(document.createTextNode(part));
-              }
-              if (i < parts.length - 1) {
-                ref.current!.appendChild(document.createElement('br'));
-              }
-            });
+            ref.current.appendChild(createTextFragment(segment.value));
           } else if (segment.type === 'badge') {
             ref.current.appendChild(createCommandBadgeElement(segment.data));
           }
